Clean up -Werror=uninitialized in curve public parameters.
Default constructor of all our curves returns the corresponding point
at infinity via a code similar to:

G()
{
   this->X = G::G_zero.X;
   this->Y = G::G_zero.Y;
   this->Z = G::G_zero.Z;
}

This gives a compile-time warning as, of course, G_zero is a complex
object (containing field elements in Montgomery form, for example) and
has not been initialized (as init_public_params() is meant to be
called at runtime).

This commit suppresses this warning by introducing a sentinel variable
G::initialized which we statically initialize to false, and set to
true in init_public_params(). We then wrap code of the default
constructor in a "if (G::initialized) { ... }" block.

Note that this commit does _not_ introduce additional error checking
for the scenarion when init_public_params is not called (no
operations, apart from the default constructor access ::initialized).
It is still user's responsibility to use init_public_params.